Add a common AbstractCommand::parseNumber() function
This is supersedes:
* parseMicroValue()
* parseMilliValue
* parseWholeValue()

Those functions are now just gutted wrappers of parseNumber().

This also we can have a parseNanoValue() equivalent, which we need
for Pokit Pro's capacitance mode.

Those parse*Value() functions will be removed next, but I've kept
them in this commit to test/prove that all the existing unit tests
still pass on all CI platforms first.