Remove some old support for CMake versions pre-3.12
Specifically, we use the target_link_libraries command's support
for Object Libraries (added in CMake 3.12) so our unit tests can
link against the same objects as the standalone apps whose code is
being tested.  We had a fallback for pre-3.12, which allowed to
build but disabled unit tests.  This is no longer relevabt since
we're required CMake 3.15+ now for other reasons.