Ignore inconsistencies with genhtml too
Required for LCOV v2.3+

Note, these inconsistencies arise because Clang and GCC
instrumentation differs in their record of which lines some
function begin and end, especially when the functions' openning
`{` is on the following line from the function declartion.

Hopefully the two will agree one day, but I very much doubt it.

Possibly the `lcov` and `genhtml` commands will reconcile, but I
wouldn't expect that to be any time soon.

Also added `x86-64` to the `removeHtmlDates.sh` script path,
because since Linux arm64 support in #d54fbbc, the glob was
matching two instances, which the Bash invocation was silently
ignoring.